Debugging linux organ pipe filter

+++ Main/extras/MKAiff.c
@@ -1684,7 +1684,7 @@ int aiffPlayheadPositionInFrames(MKAiff* aiff)
 //playheadPositionInSamples--------------------------------------------------------------
 int aiffPlayheadPositionInSamples(MKAiff* aiff)
 {
-  uint64_t offset = (uint64_t)aiff->playhead - (uint64_t)aiff->soundBuffer;
+  uint64_t offset = aiff->playhead - aiff->soundBuffer;
   offset /= 4;
   //printf("offset: %u, bufferSize: %u\n", (unsigned int)offset, (unsigned int)aiff->bufferCapacityInSamples);
   return offset;
